Fix and enhancement
